requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
async-timeout==4.0.3
    # via redis
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via soundfile
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   gtts
    #   rq
ffmpeg-python==0.2.0
    # via openai-whisper
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
future==1.0.0
    # via ffmpeg-python
greenlet==3.0.3
    # via sqlalchemy
gtts==2.5.2
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   ollama
    #   python-telegram-bot
huggingface-hub==0.24.2
    # via
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
loguru==0.7.2
    # via -r requirements.in
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
more-itertools==10.3.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==2.0.1
    # via
    #   -r requirements.in
    #   numba
    #   openai-whisper
    #   transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.5.82
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
ollama==0.3.0
    # via -r requirements.in
openai-whisper==20230306
    # via -r requirements.in
packaging==24.1
    # via
    #   huggingface-hub
    #   transformers
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   sqlmodel
pydantic-core==2.20.1
    # via pydantic
pydub==0.25.1
    # via -r requirements.in
python-telegram-bot==21.4
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
redis==5.0.7
    # via
    #   -r requirements.in
    #   rq
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   gtts
    #   huggingface-hub
    #   speechrecognition
    #   transformers
rq==1.16.2
    # via -r requirements.in
safetensors==0.4.3
    # via transformers
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soundfile==0.12.1
    # via -r requirements.in
speechrecognition==3.10.4
    # via -r requirements.in
sqlalchemy==2.0.31
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.21
    # via -r requirements.in
sympy==1.13.1
    # via torch
tokenizers==0.19.1
    # via transformers
torch==2.4.0
    # via
    #   -r requirements.in
    #   openai-whisper
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   openai-whisper
    #   transformers
transformers==4.43.2
    # via openai-whisper
triton==3.0.0
    # via
    #   openai-whisper
    #   torch
typing-extensions==4.12.2
    # via
    #   alembic
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   speechrecognition
    #   sqlalchemy
    #   torch
urllib3==2.2.2
    # via requests
